[refactor] added config files
