+++ src/configs/cache-query-commands.config/cache-query-commands.config.module.ts
@@ -1,0 +1,4 @@
+import { Module } from '@nestjs/common';
+
+@Module({})
+export class CacheQueryCommandsConfigModule {}
