+++ src/configs/cache-query-commands.config/cache-query-commands.config.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class CacheQueryCommandsConfigService {}
