+++ src/configs/t-line-calculator.config/t-line-calculator.config.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { TLineCalculatorConfigService } from './t-line-calculator.config.service';
+
+@Module({
+  providers: [TLineCalculatorConfigService],
+  exports: [TLineCalculatorConfigService]
+})
+export class TLineCalculatorConfigModule {}
