+++ src/configs/t-line-calculator.config/t-line-calculator.config.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class TLineCalculatorConfigService {}
