+++ src/modules/post-ranker/post-ranker.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PostRankerService } from './post-ranker.service';
-import { NeoQueryModule } from '../../neo-query/neo-query.module';
+import { NeoQueryModule } from '../neo-query/neo-query.module';
 import { TlineCacherModule } from '../tline-cacher/tline-cacher.module';
 
 @Module({
